+++ src/Vortice.Direct2D1/CustomVertexBufferProperties.cs
@@ -2,7 +2,6 @@
 // Distributed under the MIT license. See the LICENSE file in the project root for more information.
 
 using System;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 namespace Vortice.Direct2D1