+++ src/Vortice.Direct2D1/DirectWrite/IDWriteLocalizedStrings.cs
@@ -1,10 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿// Copyright (c) Amer Koleci and contributors.
+// Distributed under the MIT license. See the LICENSE file in the project root for more information.
+
+using System;
 
 namespace Vortice.DirectWrite
 {
-    partial class IDWriteLocalizedStrings
+    public partial class IDWriteLocalizedStrings
     {
         public unsafe string GetString(int index)
         {
@@ -13,6 +14,7 @@ namespace Vortice.DirectWrite
             GetString(index, new IntPtr(chars), length + 1);
             return new string(chars, 0, length);
         }
+
         public unsafe string GetLocaleName(int index)
         {
             var length = GetLocaleNameLength(index);