+++ src/Vortice.Direct2D1/DirectWrite/IDWriteTextRenderer1.cs
@@ -1,6 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿// Copyright (c) Amer Koleci and contributors.
+// Distributed under the MIT license. See the LICENSE file in the project root for more information.
+
+using System;
 using SharpGen.Runtime;
 using Vortice.Direct2D1;