+++ src/Vortice.Direct2D1/DirectWrite/IDWriteTextRenderer1Shadow.cs
@@ -1,7 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
+﻿// Copyright (c) Amer Koleci and contributors.
+// Distributed under the MIT license. See the LICENSE file in the project root for more information.
+
+using System;
 using System.Runtime.InteropServices;
-using System.Text;
 using SharpGen.Runtime;
 using SharpGen.Runtime.Win32;
 using Vortice.Direct2D1;