+++ src/Vortice.Direct2D1/DirectWrite/IDWriteTextRendererShadow.cs
@@ -1,8 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Runtime.CompilerServices;
+﻿// Copyright (c) Amer Koleci and contributors.
+// Distributed under the MIT license. See the LICENSE file in the project root for more information.
+
+using System;
 using System.Runtime.InteropServices;
-using System.Text;
 using SharpGen.Runtime;
 using Vortice.Direct2D1;