+++ src/Vortice.Direct2D1/ID2D1Transform.cs
@@ -1,6 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿// Copyright (c) Amer Koleci and contributors.
+// Distributed under the MIT license. See the LICENSE file in the project root for more information.
+
 using SharpGen.Runtime;
 using Vortice.Mathematics;
 
@@ -14,5 +14,4 @@ namespace Vortice.Direct2D1
         void MapInputRectsToOutputRect(Rect[] inputRects, Rect[] inputOpaqueSubRects, out Rect outputRect, out Rect outputOpaqueSubRect);
         void MapInvalidRect(int inputIndex, Rect invalidInputRect, out Rect invalidOutputRect);
     }
-    
 }