+++ src/Vortice.Direct2D1/WIC/IWICPalette.cs
@@ -1,6 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿// Copyright (c) Amer Koleci and contributors.
+// Distributed under the MIT license. See the LICENSE file in the project root for more information.
 
 namespace Vortice.WIC
 {