+++ src/Vortice.Multimedia/WaveFormatEncoding.cs
@@ -1,11 +1,6 @@
 // Copyright (c) Amer Koleci and contributors.
 // Distributed under the MIT license. See the LICENSE file in the project root for more information.
 
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Runtime.InteropServices;
-
 namespace Vortice.Multimedia
 {
     /// <summary>